make testRules a util function
